requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.18.0
fsspec==2025.3.0
h11==0.14.0
huggingface-hub==0.14.1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
transformers==4.30.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
selenium==4.30.0
webdriver-manager==4.0.2
openai==1.68.2
python-dotenv==1.0.1
